# backendPI4/api/openapi.yaml
openapi: 3.0.0
info:
  title: API para a aplicação ViriatosScouting
  description:
    "Este documento apresenta toda a documentação de uma API que permite\
    \ gerir um sistema que faz gestão de uma equipa de scouting. A API inclui funcionalidades\
    \ que permitem o utilizador gerir jogadores, clubes, escalões, equipas e utilizadores.\
    \ A mesma tambem apresenta um local para criar tarefas e relatórios."
  version: "1.0"
servers:
  - url: https://localhost:8080
    description: Servidor local utilizado para testar a API
  - url: https://backendpi4-bbzf.onrender.com
    description: Servidor online utilizado para testar a API
...
paths:
  /atletas:
    get:
      tags:
        - Atletas
      summary: Lista todos os atletas
      description:
        "Retorna uma lista de todos os atletas, mas somente administradores\
        \ e outro podem ver esta informação. Para obter os atletas, o utilizador deve\
        \ estar autenticado com um token válido. O mesmo tambem já faz a filtragem\
        \ por ano de nascimento, nacionalidade, posição, avaliação final, clube e\
        \ escalão."
      operationId: listaAtletas
      parameters:
        - name: anoNascimento
          in: query
          description: Ano de nascimento para filtrar os atletas.
          required: false
          style: form
          explode: true
          schema:
            maximum: 2100
            minimum: 1900
            type: integer
            format: int32
            example: 1985
        - name: nacionalidade
          in: query
          description: Nacionalidade para filtrar os atletas.
          required: false
          style: form
          explode: true
          schema:
            maxLength: 512
            pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
            type: string
            example: Português
        - name: posicao
          in: query
          description: Posição para filtrar os atletas.
          required: false
          style: form
          explode: true
          schema:
            maximum: 99
            minimum: 1
            type: integer
            format: int32
            example: 4
        - name: avaliacaofinal
          in: query
          description: Avaliação final para filtrar os atletas.
          required: false
          style: form
          explode: true
          schema:
            maximum: 5
            minimum: 1
            type: integer
            format: int32
            example: 5
        - name: clube_id
          in: query
          description: ID do clube para filtrar os atletas.
          required: false
          style: form
          explode: true
          schema:
            maximum: 999999
            minimum: 1
            type: integer
            format: int64
            example: 1
        - name: escalao_id
          in: query
          description: ID do escalão para filtrar os atletas.
          required: false
          style: form
          explode: true
          schema:
            maximum: 999999
            minimum: 1
            type: integer
            format: int64
            example: 2
        - name: limit
          in: query
          description: Número máximo de atletas a serem retornadas por página.
          required: false
          style: form
          explode: true
          schema:
            maximum: 999999
            minimum: 1
            type: integer
            format: int64
            example: 10
            default: 10
        - name: offset
          in: query
          description: Número de atletas a ignorar para paginação.
          required: false
          style: form
          explode: true
          schema:
            maximum: 300
            minimum: 0
            type: integer
            format: int64
            example: 1
            default: 0
        - name: nome
          in: query
          description: Nome do atleta a ir buscar.
          required: false
          style: form
          explode: true
          schema:
            maximum: 30
            minimum: 0
            type: string
            example: "Cristiano Ronaldo"
      responses:
        "200":
          description: Lista de atletas com metadados de paginação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListarAtleta200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma atleta encontrado para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorAtleta"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Atletas
    post:
      tags:
        - Atletas
      summary: Cria um novo atleta
      description:
        Permite criar um novo atleta. Apenas administradores e olheiros
        podem criar atletas. O utilizador deve fornecer os dados do atleta no body.
      operationId: criarAtleta
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtletaCriar"
        required: true
      responses:
        "201":
          description: Atleta criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtletaCriar"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Atletas
  /atletas/{id}:
    get:
      tags:
        - Atletas
      summary: Lista um atleta
      description: "Retorna uma lista de um atleta, mas somente administradores\
        \ e outro podem ver esta informação. Para obter os atletas, o utilizador deve\
        \ estar autenticado com um token válido. O mesmo tambem já faz a filtragem\
        \ por ano de nascimento, nacionalidade, posição, avaliação final, clube e\
        \ escalão."
      operationId: listaUmAtleta
      parameters:
        - name: id
          in: path
          description: ID do atleta a mostrar
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 100
            pattern: "^[0-9]+$"
            type: string
      responses:
        "200":
          description: Lista de atletas com metadados de paginação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListarUmAtleta200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma atleta encontrado para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorAtleta"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Atletas
    put:
      tags:
        - Atletas
      summary: Edita um atleta
      description:
        Permite editar um atleta existente. O utilizador pode editar atletas
        se o mesmo for administrador. É necessário um token válido para autenticação.
      operationId: editarAtleta
      parameters:
        - name: id
          in: path
          description: ID do atleta a ser atualizado
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 100
            pattern: "^[0-9]+$"
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtletaEditar"
        required: true
      responses:
        "200":
          description: Atleta atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtletaEditar"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Atletas
    delete:
      tags:
        - Atletas
      summary: Apaga um atleta
      description:
        Permite apagar um atleta existente. Apenas o administrador pode
        apagá-lo. O utilizador deve fornecer o ID do atleta a ser apagado.
      operationId: apagarAtleta
      parameters:
        - name: id
          in: path
          description: ID do atleta a ser apagado
          required: true
          style: simple
          explode: false
          schema:
            maxLength: 100
            pattern: "^[0-9]+$"
            type: string
      responses:
        "204":
          description: Atleta apagado com sucesso
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhum atleta encontrado para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorAtleta"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Atletas
  /atletas/todos:
    get:
      tags:
        - Atletas
      summary: Lista todos os atletas
      description:
          "Retorna uma lista de todos os atletas, mas somente administradores\
          \ e outro podem ver esta informação. Para obter os atletas, o utilizador deve\
          \ estar autenticado com um token válido. O mesmo tambem já faz a filtragem\
          \ por ano de nascimento, nacionalidade, posição, avaliação final, clube e\
          \ escalão."
      operationId: listartodosatletas
      responses:
        "200":
          description: Lista de atletas com metadados de paginação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListarAtleta200"
        "400":
          description: Dados inválidos - Erro no formato ou campos obrigatórios ausentes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestError"
        "401":
          description: Não autorizado - Token de autenticação inválido ou ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: Nenhuma atleta encontrado para os filtros aplicados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorAtleta"
        "500":
          description: Erro inesperado - Ocorreu um erro inesperado no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnexpectedError"
      security:
        - bearerAuth: []
      x-swagger-router-controller: Atletas
...
components:
  schemas:
    Atleta:
      required:
        - anoNascimento
        - clube_id
        - escalao_id
        - id
        - nacionalidade
        - nome
        - posicao
      type: object
      properties:
        id:
          maximum: 999999
          minimum: 1
          type: integer
          description: Identificador único do atleta
          format: int64
          example: 1
        nome:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nome completo do atleta
          example: Cristiano Ronaldo
        dataNascimento:
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          type: string
          description: Data de nascimento do atleta (formato YYYY-MM-DD)
          format: date
        anoNascimento:
          maximum: 2100
          minimum: 1900
          type: integer
          description: Ano de nascimento do atleta
          format: int32
          example: 1985
        nacionalidade:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nacionalidade do atleta´
          example: Português
        posicao:
          maximum: 4
          minimum: 1
          type: integer
          description: Identificador numérico para a posição do atleta
          format: int32
          example: 4
        links:
          type: array
          description: Lista de links relacionados ao atleta
          items:
            $ref: "#/components/schemas/Link"
        nomeAgente:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nome do agente
          example: Jorge Mendes
        numeroAgente:
          maxLength: 512
          pattern: "^\\+351\\s\\d{9}$"
          type: string
          description: Contacto do agente
          example: +351 912345678
        avaliacaoFinal:
          maximum: 5
          minimum: 1
          type: integer
          description: Avaliação final do atleta (1-5)
          format: int32
          example: 5
        clube_id:
          $ref: "#/components/schemas/Clube"
        escalao_id:
          $ref: "#/components/schemas/EscalaoAtleta"
      additionalProperties: false
    AtletaCriar:
      required:
        - anoNascimento
        - nacionalidade
        - nome
      type: object
      properties:
        nome:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nome completo do atleta
          example: Cristiano Ronaldo
        dataNascimento:
          type: string
          format: date
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          description: Data de nascimento do atleta (formato YYYY-MM-DD)
          nullable: true
        anoNascimento:
          maximum: 2100
          minimum: 1900
          type: integer
          description: Ano de nascimento do atleta
          format: int32
          example: 1985
        nacionalidade:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nacionalidade do atleta
          example: Português
        posicao:
          type: integer
          description: Identificador numérico para a posição do atleta
          example: 4
          nullable: true
        nomeAgente:
          type: string
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          description: Nome do agente
          example: Jorge Mendes
          nullable: true
        numeroAgente:
          type: string
          maxLength: 512
          pattern: "^\\+351\\s\\d{9}$"
          description: Contacto do agente
          example: "+351 912345678"
          nullable: true
        links:
          maxItems: 10
          pattern:
            "^(https?|ftp):\\/\\/[\\w\\-]+(\\.[\\w\\-]+)+(\\/[\\w\\-\\.~:@\\\
            /?#\\[\\]@!\\$&'\\(\\)\\*\\+,;=%]*)?$"
          type: array
          description: Lista de id de links relacionados ao atleta
          items:
            maxLength: 512
            type: string
            example: https://www.zerozero.pt/jogador/gil-graca/528406
          nullable: true
        idClube:
          type: integer
          description: Identificador único do clube
          example: 1
          nullable: true
        idEscalao:
          type: integer
          description: Identificador único do escalão
          example: 1
          nullable: true
      additionalProperties: false
    AtletaEditar:
      required:
        - anoNascimento
        - nacionalidade
        - nome
      type: object
      properties:
        nome:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nome completo do atleta
          example: Cristiano Ronaldo
        dataNascimento:
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          type: string
          description: Data de nascimento do atleta (formato YYYY-MM-DD)
          format: date
          nullable: true
        anoNascimento:
          maximum: 2100
          minimum: 1900
          type: integer
          description: Ano de nascimento do atleta
          format: int32
          example: 1985
        nacionalidade:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nacionalidade do atleta´
          example: Português
        posicao:
          maximum: 4
          minimum: 1
          type: integer
          description: Identificador numérico para a posição do atleta
          format: int32
          example: 4
          nullable: true
        nomeAgente:
          maxLength: 512
          pattern: "^[a-zA-ZÀ-ÿ'\\- ]+$"
          type: string
          description: Nome do agente
          example: Jorge Mendes
          nullable: true
        numeroAgente:
          maxLength: 512
          pattern: "^\\+351\\s\\d{9}$"
          type: string
          description: Contacto do agente
          example: +351 912345678
          nullable: true
        links:
          type: array
          description: Lista de links relacionados ao atleta
          maxItems: 10
          items:
            type: object
            properties:
              link:
                type: string
                description: URL relacionada ao atleta
                maxLength: 512
                example: https://www.zerozero.pt/jogador/gil-graca/528406
                nullable: true
            required:
              - link
          pattern:
            "^(https?|ftp):\\/\\/[\\w\\-]+(\\.[\\w\\-]+)+(\\/[\\w\\-\\.~:@\\/\\?#\\[\\]@!\\$&'\\(\\)\\*\\+,;=%]*)?$"
        avaliacaoFinal:
          maximum: 5
          minimum: 1
          type: integer
          description: Avaliação final do atleta (1-5)
          format: int32
          nullable: true
          example: 5
        idClube:
          maximum: 999999
          minimum: 1
          type: integer
          description: Identificador único do atleta
          format: int64
          example: 1
          nullable: true
        idEscalao:
          maximum: 999999
          minimum: 1
          type: integer
          description: Identificador único do atleta
          format: int64
          example: 1
          nullable: true
      additionalProperties: false
...
    EscalaoAtleta:
      required:
        - nome
      type: object
      properties:
        nome:
          maxLength: 512
          type: string
          description: Nome do escalão
          example: Juvenis
      additionalProperties: false
...
    UnexpectedError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9À-ÿ ]+$"
          type: string
          example: Ocorreu um erro inesperado no servidor.
      additionalProperties: false
    BadRequestError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          example: Token de autenticação inválido ou ausente.
      additionalProperties: false
    UnauthorizedError:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          example: Campos obrigatórios ausentes ou formato inválido.
      additionalProperties: false
...
    NotFoundErrorAtleta:
      type: object
      properties:
        error:
          maxLength: 255
          pattern: "^[A-Za-z0-9 ]+$"
          type: string
          description: Mensagem explicativa do erro
          example: Nenhum atleta encontrado.
      additionalProperties: false
...
    ListarAtleta200:
      type: object
      properties:
        totalCount:
          type: integer
          description: Número total de atletas que correspondem aos filtros aplicados.
        totalPages:
          type: integer
          description: "Número total de páginas, baseado no 'limit' fornecido."
        currentPage:
          type: integer
          description: Página atual (calculada com o offset e limit).
        data:
          maxItems: 50
          type: array
          items:
            $ref: "#/components/schemas/Atleta"
      additionalProperties: false
    ListarUmAtleta200:
      type: object
      properties:
        data:
          maxItems: 50
          type: array
          items:
            $ref: "#/components/schemas/Atleta"
      additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT